make auth requests from frontend

+++ frontend/src/components/WorkoutDetails.js
@@ -1,14 +1,22 @@
 import React from 'react'
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
+import { useAuthContext } from '../hooks/useAuthContext'
 
 export default function WorkoutDetails({workout}) {
 
     const {dispatch} = useWorkoutsContext()
+    const {user} =useAuthContext()
 
    const handleClick = async ()=>{
+
+    if(!user) return
+
     const response = await fetch('/api/workouts/'+workout._id,{
-      method:'DELETE'
+      method:'DELETE',
+      headers:{
+        'Authorization':`Bearer ${user.token}`
+      }
     })
     console.log('fired')
